Clarify the docstring for alignment(::PairwiseAlignmentResult)
alignment(::AlignedSequence) and alignment(::PairwiseAlignmentResult)
both return an Alignment, but alignment(::PairwiseAlignmentResult) has
always returned a PairwiseAlignment. I don't want to break the current
implementation for alignment(::PairwiseAlignmentResult), but the the
implementation of the two others makes sense in the context of their types.
Ideally, I would separate the alignment(::PairwiseAlignmentResult) function
into separate getter functions like
alignment(::PairwiseAlignmentResult)::Alignment and
pairwise_alignment(::PairwiseAlignmentResult)::PairwiseAlignment, but
again, that would be breaking. So instead, settle for a really
dirty-feeling compromise of documenting the difference and explaining the
reasoning behind it with the hope that BioAlignments v4 is able to
reconcile this bandaged solution.